feat: pull out react query hooks into own file
