+++ app/frontend/src/features/connections/friends/FriendRequestsReceived.tsx
@@ -1,49 +1,13 @@
 import { Box, IconButton } from "@material-ui/core";
-import { Empty } from "google-protobuf/google/protobuf/empty_pb";
-import { Error } from "grpc-web";
 import React from "react";
-import { useMutation, useQueryClient } from "react-query";
 import { CheckIcon, CloseIcon } from "../../../components/Icons";
 import { FriendRequest } from "../../../pb/api_pb";
-import { service } from "../../../service";
 import FriendSummaryView from "./FriendSummaryView";
 import FriendTile from "./FriendTile";
 import useFriendRequests from "./useFriendRequests";
+import useRespondToFriendRequest from "./useRespondToFriendRequest";
 import { useIsMounted, useSafeState } from "../../../utils/hooks";
 import type { SetMutationError } from ".";
-
-interface RespondToFriendRequestVariables {
-  accept: boolean;
-  friendRequestId: number;
-  setMutationError: SetMutationError;
-}
-
-export function useRespondToFriendRequest() {
-  const queryClient = useQueryClient();
-  const { mutate: respondToFriendRequest } = useMutation<
-    Empty,
-    Error,
-    RespondToFriendRequestVariables
-  >(
-    ({ friendRequestId, accept }) =>
-      service.api.respondFriendRequest(friendRequestId, accept),
-    {
-      onMutate: async ({ setMutationError }) => {
-        setMutationError("");
-        await queryClient.cancelQueries("friendRequestsReceived");
-      },
-      onSuccess: () => {
-        queryClient.invalidateQueries("friendIds");
-        queryClient.invalidateQueries("friendRequestsReceived");
-      },
-      onError: (error, { setMutationError }) => {
-        setMutationError(error.message);
-      },
-    }
-  );
-
-  return respondToFriendRequest;
-}
 
 interface RespondToFriendRequestActionProps {
   friendRequestId: number;
@@ -107,7 +71,7 @@ function FriendRequestsReceived() {
         data.map((friendRequest) => (
           <FriendSummaryView
             key={friendRequest.friendRequestId}
-            friendRequest={friendRequest}
+            friend={friendRequest.friend}
           >
             <RespondToFriendRequestAction
               friendRequestId={friendRequest.friendRequestId}
